% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdens-to-points.R
\name{pop2point}
\alias{pop2point}
\title{Assign raster cell population density values to points}
\usage{
pop2point(verts, geotiff, normalise = TRUE)
}
\arguments{
\item{verts}{A table of vertices with longitude and latitude coordinates,
such as the "verts" component of a \pkg{silicate}-class object representing a
local street network, generated with \pkg{dodgr} function,
\code{dodgr_streetnet_sc()}.}

\item{geotiff}{Path to 'geotiff' file containing population density
estimates, and including the area defined by 'net_sc'.}

\item{normalise}{If \code{TRUE}, normalise results so such of densities at all
network vertices is equal to sum of original densities.}
}
\value{
A \code{data.frame} containing all vertices of 'net_sc', and
corresponding point estimates of population density.
}
\description{
Assign raster cell population density values to points
}
