% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-binary-compare.R
\name{ua_binary_compare}
\alias{ua_binary_compare}
\title{Statistically compare UA values between two spatial groups}
\usage{
ua_binary_compare(city, results_path, soc, bin_var, d = 10)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{results_path}{Local path to directory where results are to be written.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{bin_var}{The same of the variable in "soc" used to distinguish binary
groups. This must be a logical variable (values as \code{TRUE} or \code{FALSE} only).}

\item{d}{One value of 'dlim' parameters used in \link{ua_index} or
\link{ua_index_batch} call to generate UA data. Resultant plot will be
based on data for this value.}
}
\description{
Spatial groups must be distinguished by a named variable in the "soc"
data.table.
}
