% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-batch.R
\name{ua_index_batch}
\alias{ua_index_batch}
\title{Batch calculation of 'ua_index' for a single city}
\usage{
ua_index_batch(
  city,
  gtfs_path,
  popdens_geotif = NULL,
  results_path = NULL,
  initial_mode = "foot",
  final_mode = "foot",
  soc = NULL,
  soc_var = NULL,
  dlims = c(5, 10),
  batch_size = 1000,
  coverage = 1/4,
  nthreads = NULL
)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{gtfs_path}{Path to \code{.Rds} version of GTFS feed for specified city.}

\item{popdens_geotif}{Optional path to local 'geotiff' file with population
density estimates. If provided, all transport indices are adjusted to account
for effects of local population density.}

\item{results_path}{Local path to directory where results are to be written.}

\item{initial_mode}{Initial mode of transport from each 'from' point to
public transport system (or to destination points, where single-model
transport is faster).}

\item{final_mode}{Final mode of transport from each public transport
destination point to all other network nodes.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{soc_var}{Name of target variable in \code{soc} data set.}

\item{dlims}{Vector of distance limits in kilometres over which average
values of transport index should be calculated. One value of "ua_index" is
then derived for each value of \code{dlims}.}

\item{batch_size}{Number of vertices to calculate in each batch before saving
to local file. Increases in this parameter can greatly increase the sizes of
intermediate results, leading to memory problems. Note that batch runs will
only be able to stopped and restarted if this parameter is kept constant.
Changes to this parameter will cause re-started runs to revert to the
beginning, and will generate results which overlap with any previous results
still retained in \code{results_path}.}

\item{coverage}{Proportion of all points for which values are to be
calculated. As described above, coverage can generally safely be set to
values less than 1, with complete coverage then generated through the more
computationally efficient \link{ua_interpolate} function.}

\item{nthreads}{Optional parameter to specify number of threads to be used in
multi-threaded calculations. Default is maximum number of available threads
(from \code{RcppParallel::defaultNumThreads()}).}
}
\description{
Repeatedly calculates values from the \link{ua_index} function
for a sub-set of all origin vertices within nominated city, and saves results
to a local file. Finally results can generally be reliably interpolated, and
it is thus generally not necessarily to calculate \link{ua_index} for every
origin point. Instead, it is recommended to only calculate values for a fixed
proportion specified by the parameter, \code{coverage}.
}
\note{
Importantly, this function can safely be interrupted at any time, and
will simply restart at the point where it was stopped.
}
