% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-interpolate.R
\name{ua_interpolate}
\alias{ua_interpolate}
\title{Interpolate "UA_index" between vertices on network, project those on to the
original network edges, and return the full network graph.}
\usage{
ua_interpolate(city, initial_mode = "foot", ua_dat = NULL, soc, npts = 3L)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{initial_mode}{Initial mode of transport from each 'from' point to
public transport system (or to destination points, where single-model
transport is faster).}

\item{ua_dat}{Result of \link{ua_index} function.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable, as passed to \link{ua_index}
function.}

\item{npts}{Number of nearest neighbours to use to interpolate values at each
point.}
}
\value{
The full network graph with additional variables quantifying for each
edges the values defined in \code{ua_dat} as projected on to those edges.
}
\description{
Interpolate "UA_index" between vertices on network, project those on to the
original network edges, and return the full network graph.
}
