% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-plot.R
\name{ua_plot_network}
\alias{ua_plot_network}
\title{Generate an interactive plot of results from 'ua_interpolate' projected on
to street network.}
\usage{
ua_plot_network(graph, var = "ua_index_d10")
}
\arguments{
\item{graph}{A street network with values of UA index calculated from
\link{ua_index} and interpolated back onto network with
\link{ua_interpolate}.}

\item{var}{Which variable to plot.}
}
\value{
Nothing; function called for its side-effect of opening an
interactive visualisation in local default browser.
}
\description{
Generate an interactive plot of results from 'ua_interpolate' projected on
to street network.
}
