% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-to-geojson.R
\name{ua_to_geojson}
\alias{ua_to_geojson}
\title{Write UA index data to local 'geojson' file}
\usage{
ua_to_geojson(ua_dat, ndigits = 2L, filename = NULL)
}
\arguments{
\item{ua_dat}{Full network graph including UA data returned from
\link{ua_index} and \link{ua_interpolate} functions.}

\item{ndigits}{Number of digits for rounding UA data in 'geojson'
representation.}

\item{filename}{Name and/or full path to local file where 'geojson' data are
to be stored.}
}
\description{
Writing large graphs to 'geojson' output can take some time - typically on
the order of minutes per 100,000 graph edges.
}
