% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices.R
\name{ua_vertices}
\alias{ua_vertices}
\title{Extract vertices within perimiter of external data set}
\usage{
ua_vertices(x, city, mode)
}
\arguments{
\item{x}{External data set, presumed to represent some socio-demographic
variable, containing values of defined variable and \pkg{sf}-format polygons
defining each observation of that variable.}

\item{city}{Name of city for which vertices are extracted; passed to
\pkg{m4ra} functions which store pre-processed networks in locations defined
by name of 'city'}

\item{mode}{Mode of transport used to generate network; this should be the
same as the 'initial_mode' parameter passed to \pkg{m4ra} functions.}
}
\value{
A vertex table of only those vertices within the bounding polygon
defined by \code{x}.
}
\description{
Extract vertices within perimiter of external data set
}
