% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uaengine-package.R
\docType{package}
\name{uaengine-package}
\alias{uaengine}
\alias{uaengine-package}
\title{uaengine: Routing and aggregation engine for 'Urban Analyst'}
\description{
Routing and aggregation engine for 'Urban Transport Analyst'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/uaengine}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/uaengine/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com}
}

}
\keyword{internal}
