% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-binary-compare.R
\name{uta_binary_compare}
\alias{uta_binary_compare}
\title{Statistically compare UTA values between two spatial groups}
\usage{
uta_binary_compare(city, results_path, soc, bin_var, d = 10)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{results_path}{Local path to directory where results are to be written.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}
}
\description{
Spatial groups must be distinguished by a named variable in the "soc"
data.table.
}
