% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-batch.R
\name{uta_export}
\alias{uta_export}
\title{Export results to the form required for the front-end app}
\usage{
uta_export(city, soc, results_path, popdens_geotif)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{soc}{An \pkg{sf} \code{data.frame} object with polygons defining areas in
which socio-demographic variables were collated, and a column called
"social_index" containing the socio-demographic variable of interest.}

\item{results_path}{Local path to directory where results are to be written.}

\item{popdens_geotif}{Optional path to local 'geotiff' file with population
density estimates. If provided, all transport indices are adjusted to account
for effects of local population density.}
}
\value{
The input table, \code{soc}, with additional columns of UTA index values
attached. This table can then be saved and used directly in the UTA
front-end.
}
\description{
Export results to the form required for the front-end app
}
