% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmium-extract.R
\name{uta_extract_osm}
\alias{uta_extract_osm}
\title{Extract network data with osmium, convert to \pkg{sc} format, and collate all
results into a single \pkg{osmdata} object.}
\usage{
uta_extract_osm(city, path_to_bz2, bbox = NULL, bbox_expand = 0.05)
}
\arguments{
\item{path_to_bz2}{Path to a local \code{.bz2} file with OpenStreetMap data for
nominated city.}

\item{bbox}{Optional bounding box within which data are to be extracted. If
not given, result includes the entire network within the nominated '.bz2'
file. \code{bbox} can be either a matrix obtained from the \pkg{osmdata} function,
\code{getbb} (or equivalent), or an object from which a bounding box can be
extracted. Objects currently recognised are matrices or arrays, which should
have two columns (x and y / longitude and latitude), or an \pkg{sf} object
from which a bounding box can be extracted.}

\item{bbox_expand}{A proportional amount by which to extend the limits of the
bounding box defined by the \code{bbox} argument, defaulting to 5\%.}

\item{City}{Name of city (used to name resultant files).}
}
\description{
Extract network data with osmium, convert to \pkg{sc} format, and collate all
results into a single \pkg{osmdata} object.
}
