% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-plot.R
\name{uta_plot}
\alias{uta_plot}
\title{Generate an interactive plot of results from 'uta_interpolate'}
\usage{
uta_plot(graph, var = "uta_index_d10")
}
\arguments{
\item{graph}{A street network with values of UTA index calculated from
\link{uta_index} and interpolated back onto network with
\link{uta_interpolate}.}

\item{var}{Which variable to plot.}
}
\value{
Nothing; function called for its side-effect of opening an
interactive visualisation in local default browser.
}
\description{
Generate an interactive plot of results from 'uta_interpolate'
}
