% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-plot.R
\name{uta_plot_polygons}
\alias{uta_plot_polygons}
\title{Generate an interactive plot of results in same polygons as original
socio-demographic data.}
\usage{
uta_plot_polygons(
  city,
  results_path,
  soc,
  d = 10,
  what = c("transport", "social", "uta_rel", "uta_abs"),
  zoom = 10,
  alpha = 0.5
)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{results_path}{Local path to directory where results are to be written.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{d}{One value of 'dlim' parameters used in \link{uta_index} or
\link{uta_index_batch} call to generate UTA data. Resultant plot will be
based on data for this value.}

\item{what}{The value which is to be plotted, one of "transport", "social",
"uta_rel" or "uta_abs". The relative UTA index ("uta_rel") combines
socio-demographic variable, "soc", with the transport index relative to
equivalent automobile travel times. The absolute UTA index ("uta_abs")
combines uses as a transport index the absolute multi-modal travel times for
all modes excluding private automobile.}

\item{zoom}{Initial zoom level to use in resultant map.}

\item{alpha}{Transparency of polygons (0 = completely; 1 = not transparent).}
}
\value{
Nothing; function called for its side-effect of opening an
interactive visualisation in local default browser.
}
\description{
Generate an interactive plot of results in same polygons as original
socio-demographic data.
}
