% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{uta_prepare_data}
\alias{uta_prepare_data}
\title{Additional data preparation beyond standard \code{m4ra} routines}
\usage{
uta_prepare_data(osm_path, water_dist = 20, quiet = FALSE)
}
\arguments{
\item{osm_path}{Path to directory containing data processed by the
\link{uta_extract_osm} function for desired city.}

\item{water_dist}{Distance below which any edges are considered adjacent to
water, and thus categorised as "blue space" edges.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
Name of file holding integer index for nominated city identifying
which vertices of the foot-weighted network are within or adjacent to natural
spaces.
}
\description{
Additional data preparation beyond standard \code{m4ra} routines
}
