% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-to-geojson.R
\name{uta_to_geojson}
\alias{uta_to_geojson}
\title{Write UTA index data to local 'geojson' file}
\usage{
uta_to_geojson(uta_dat, ndigits = 2L, filename = NULL)
}
\arguments{
\item{uta_dat}{Full network graph including UTA data returned from
\link{uta_index} and {uta_interpolate} functions.}

\item{ndigits}{Number of digits for rounding UTA data in 'geojson'
representation.}

\item{filename}{Name and/or full path to local file where 'geojson' data are
to be stored.}
}
\description{
Writing large graphs to 'geojson' output can take some time - typically on
the order of minutes per 100,000 graph edges.
}
