% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utaengine-package.R
\docType{package}
\name{utaengine-package}
\alias{utaengine-package}
\alias{_PACKAGE}
\title{utaengine: Routing and aggregation engine for 'Urban Transport Analyst'}
\description{
Routing and aggregation engine for 'Urban Transport Analyst'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/uta-engine}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/uta-engine/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
