% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-batch.R
\name{uta_export}
\alias{uta_export}
\title{Export results to the form required for the front-end app}
\usage{
uta_export(city, soc, soc_var, results_path)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{soc_var}{Name of target variable in \code{soc} data set.}

\item{results_path}{Local path to directory where results are to be written.}
}
\value{
The input table, \code{soc}, with additional columns of UTA index values
attached. This table can then be saved and used directly in the UTA
front-end.
}
\description{
Export results to the form required for the front-end app
}
