% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-index.R
\name{uta_index}
\alias{uta_index}
\title{Calculate "UTA_index" for a specified set of origin locations.}
\usage{
uta_index(
  city,
  gtfs,
  popdens_geotif = NULL,
  from = NULL,
  initial_mode = "foot",
  final_mode = "foot",
  soc = NULL,
  soc_var = NULL,
  dlims = c(5, 10),
  quiet = FALSE
)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{gtfs}{Path to \code{.Rds} version of GTFS feed for specified city.}

\item{popdens_geotif}{Optional path to local 'geotiff' file with population
density estimates. If provided, all transport indices are adjusted to account
for effects of local population density.}

\item{from}{Character vector of Open Street Map IDs of vertices from which to
calculate travel times.}

\item{initial_mode}{Initial mode of transport from each 'from' point to
public transport system (or to destination points, where single-model
transport is faster).}

\item{final_mode}{Final mode of transport from each public transport
destination point to all other network nodes.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{soc_var}{Name of target variable in \code{soc} data set.}

\item{dlims}{Vector of distance limits in kilometres over which average
values of transport index should be calculated. One value of "uta_index" is
then derived for each value of \code{dlims}.}

\item{quiet}{If \code{FALSE}, display progress information on screen.}
}
\value{
An \pkg{sf} \code{data.frame}, with geometries of points defined by
\code{from}, and values for transport indices at specified distances, both raw
values and values adjusted for local population densities, along with
corresponding values of the UTA index at each value of 'dlim'.
}
\description{
Calculate "UTA_index" for a specified set of origin locations.
}
