% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-interpolate.R
\name{uta_interpolate}
\alias{uta_interpolate}
\title{Interpolate "UTA_index" between vertices on network}
\usage{
uta_interpolate(city, initial_mode = "foot", uta_dat = NULL, soc, npts = 3L)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{initial_mode}{Initial mode of transport from each 'from' point to
public transport system (or to destination points, where single-model
transport is faster).}

\item{uta_dat}{Result of \link{uta_index} function.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable, as passed to \link{uta_index}
function.}

\item{npts}{Number of nearest neighbours to use to interpolate values at each
point.}
}
\value{
A \code{data.frame} of all vertices in network, along with variables
defined in \code{uta_dat} as interpolated on to all vertices in the network.
}
\description{
Interpolate "UTA_index" between vertices on network
}
