% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uta-plot.R
\name{uta_plot_polygons}
\alias{uta_plot_polygons}
\title{Generate an interactive plot of results in same polygons as original
socio-demographic data.}
\usage{
uta_plot_polygons(
  city,
  results_path,
  soc,
  soc_var,
  d = 10,
  what = c("transport", "uta", "social"),
  zoom = 10
)
}
\arguments{
\item{city}{Name of city, used to name and define local path to
pre-calculated street networks and transport times with \pkg{m4ra} package.}

\item{results_path}{Local path to directory where results are to be written.}

\item{soc}{Socio-demographic data with an \pkg{sf}-format column of polygons
for each observed value of target variable.}

\item{soc_var}{Name of target variable in \code{soc} data set.}

\item{d}{One value of 'dlim' parameters used in \link{uta_index} or
\link{uta_index_batch} call to generate UTA data. Resultant plot will be
based on data for this value.}

\item{what}{The value which is to be plotted.}

\item{zoom}{Initial zoom level to use in resultant map.}
}
\value{
Nothing; function called for its side-effect of opening an
interactive visualisation in local default browser.
}
\description{
Generate an interactive plot of results in same polygons as original
socio-demographic data.
}
